% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmmtest.R
\name{lmmtest}
\alias{lmmtest}
\title{Testing Fixed Effects and Contrasts of the Fixed Effects}
\usage{
lmmtest(
  fit,
  index,
  contrast = NULL,
  alternative = c("two.sided", "less", "greater")
)
}
\arguments{
\item{fit}{A result of lmmfit/lmm, which contains
coef (estimates of fixed effects), a matrix with rows representing the fixed effects and columns the different response variables in the model,
cov (covariance matrix of the fixed effects), an array of three dimensions for different response variables in the model,
df (residual degree of freedom in the linear model).}

\item{index}{A vector of integers or characters indicating which fixed effects are to be tested. By default index consists of all of the fixed effects. Ignored if contrast is not NULL.}

\item{contrast}{A matrix with columns corresponding to contrasts of the fixed effects to be tested.}

\item{alternative}{A character string specifying the alternative hypothesis, one of "two.sided", "greater" or "less".}
}
\value{
A matrix of coefficients, t-values and p-values, in which the rows correspond to the features (genes) and the columns the fixed effects (covariates). .
}
\description{
lmmtest is used to test fixed effects or contrasts of fixed effects.
}
