% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm.R
\name{varest}
\alias{varest}
\title{A internal function to estimate variance components for one feature (gene).}
\usage{
varest(xxz, zrz, zryj, yryj, n, d, s, max.iter = 50, epsilon = 1e-05)
}
\arguments{
\item{xxz}{XXinv\%*\%t(ZX), where XXinv is the inverse of XX and ZX = t(Z)\%*\%X.}

\item{zrz}{ZZ - ZX\%*\%(XXinv\%*\%t(ZX)), ZZ = t(Z)\%*\%Z.}

\item{zryj}{zry[, j], where zry = ZY - ZX\%*\%(XXinv\%*\%XY)}

\item{yryj}{yry[j], where yry = Ynorm - colSums(XY*(XXinv\%*\%XY))}

\item{n}{Numbers of samples (cells in scRNA-seq).}

\item{d}{A vector of (m1,...,mk), mi = ncol(Zi), number of columns in Zi.}

\item{s}{A vector of initial values of the variance components, (s1, ...,sk, s_(k+1)).}

\item{max.iter}{The maximal number of iterations.}

\item{epsilon}{Positive convergence tolerance.}
}
\value{
A list consisting of
estimates of variance components (s),
first partial derivatives of log-likehood (dl),
number of iterations (iter), and
inverse of Fisher information matrix (Minv).
}
\description{
This function is used internally (inside lmm).
}
\keyword{internal}
